First_Project_C

+++ libft_project/ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: ylahssin <ylahssin@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/25 12:39:23 by ylahssin          #+#    #+#             */
-/*   Updated: 2024/10/25 12:44:23 by ylahssin         ###   ########.fr       */
+/*   Updated: 2024/10/29 13:27:30 by ylahssin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *ptr, int value, unsigned int num)
+#include "libft.h"
+
+void	*ft_memset(void *ptr, int value, size_t num)
 {
 	unsigned int	i;
 	unsigned char	*p;
